[audio_mic_loopback] fix bamboo errors and regen for 3.2 MH3-7607

+++ apps/microphone_loopback/firmware/src/config/sam_e70_xult_wm8904_ssc_wqvga/gfx/libaria/libaria_events.c
@@ -67,11 +67,11 @@ void _SetNumericLabel(uint32_t value, laLabelWidget* label, bool percent)
     
     if (percent)
     {
-        sprintf(valueStrBuf,"%4d%%",value);
+        sprintf(valueStrBuf,"%4d%%",(int)value);
     }
     else
     {
-        sprintf(valueStrBuf,"%4d",value);
+        sprintf(valueStrBuf,"%4d",(int)value);
     }
     tempStr = laString_CreateFromCharBuffer((char *) valueStrBuf, &LiberationSans18);
     laLabelWidget_SetText(label, tempStr);
